Make a package to use pkgimages based precompilation
