+++ julia/RelatedPosts.jl/src/vector5_staticarrays.jl
@@ -1,0 +1,4 @@
+using StaticArrays
+
+const MVector5{T} = MVector{5,T}
+const SVector5{T} = SVector{5,T}
